refactor(cli): align frontend deploy module names
